+++ src/agedi/diffusion/__init__.py
@@ -5,6 +5,7 @@ from .samplers import (
     ForcefieldCorrectorSampler,
     HeunODESampler,
     HeunSampler,
+    InpaintingSampler,
     PredictorCorrectorSampler,
     ProbabilityFlowODESampler,
     Sampler,
@@ -21,4 +22,5 @@ __all__ = [
     "ProbabilityFlowODESampler",
     "HeunODESampler",
     "ForcefieldCorrectorSampler",
+    "InpaintingSampler",
 ]